98k/wechat_sync/requirements.txt
pycryptodome>=3.20.0
zstandard>=0.22.0
pymem>=1.13.1
psutil>=5.9.8
